chore: moved jurisdiction card into folder
