+++ packages/wallets/src/features/cfd/screens/Jurisdiction/JurisdictionCard/JurisdictionCard.scss
