+++ packages/wallets/src/features/cfd/screens/Jurisdiction/JurisdictionCard/JurisdictionCardRow.scss
