+++ packages/wallets/src/features/cfd/screens/Jurisdiction/JurisdictionCard/JurisdictionCardTag.scss
