+++ packages/wallets/src/features/cfd/screens/Jurisdiction/JurisdictionCard/index.ts
@@ -1,0 +1,1 @@
+export { default as JurisdictionCard } from './JurisdictionCard';
