+++ packages/wallets/src/features/cfd/screens/Jurisdiction/JurisdictionScreen.scss
