+++ packages/wallets/src/features/cfd/screens/Jurisdiction/jurisdiction-contents/jurisdiction-svg-contents.ts
