+++ packages/wallets/src/features/cfd/screens/Jurisdiction/jurisdiction-contents/jurisdiction-vanuatu-contents.ts
