+++ packages/wallets/src/features/cfd/screens/Jurisdiction/jurisdiction-contents/jurisdiction-verification-contents.ts
