+++ packages/wallets/src/features/cfd/screens/Jurisdiction/jurisdiction-contents/jurisdiction_maltainvest_contents.ts
