+++ packages/wallets/src/features/cfd/screens/Jurisdiction/jurisdiction-contents/props.types.ts
